Move `get_running_kernel` to `PackageSack` and simplify API
